Fix heap shrinking and growing in quickly (de)allocating programs
This fixes two issues: one with the heap not growing before a recursive
GC (triggered when heap was shrunk one step before).

The other fixed issue is that the heap is now shrunk immediately (as
opposed to a "scheduled shrinking" at next GC).  In the previous
implementation it was possible that the new live set would not fit into
the (then shrunk) heap during next GC.  This now maintains the invariant
that to-space is always at least as big as from-space.