Generate a stub (runtime crash) for missing foreign functions
Creating a runtime crash (as opposed to a compile-time error) for a
missing FFI call allows to just ignore problems with functions not used
at runtime.  Previously all referenced FFI calls had to be stubbed
manually.  This makes self-hosting and keeping up with Idris2 upstream a
lot easier.

At some point in the future this should probably turned back into a
compile-time error (maybe with the possibility to ignore it via a
backend directive).