Heap hashset performance increase
The low CLUSTER_SHIFT bits of cluster start addresses are always 0, this
led to worst-case performance in the hashset used to track which memory
regions belong to our heap.

By shifting the memory address right by CLUSTER_SHIFT bits, entropy in
the lower bits of the hash key is increased. Depending on the OS/libc
implementation, entropy in the low bits of the (shifted) value might
still not be ideal. If this becomes a problem in the future, revise the
"hashing" algorithm in hashset.c (currently it's just multiplication by
73).